apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
httpx==0.27.2
python-dotenv==1.0.1
mcp[cli]==1.13.1
